+++ code/src/sixsq/nuvla/server/resources/credential.clj
@@ -12,6 +12,10 @@ passwords) or other services (e.g. TLS credentials for Docker). Creating new
     [sixsq.nuvla.server.resources.common.std-crud :as std-crud]
     [sixsq.nuvla.server.resources.common.utils :as u]
     [sixsq.nuvla.server.util.log :as logu]
+    [sixsq.nuvla.server.resources.job :as job]
+    [sixsq.nuvla.server.util.response :as r]
+    [sixsq.nuvla.server.resources.credential-template-infrastructure-service-swarm :as swarm-tpl]
+    [sixsq.nuvla.server.resources.event.utils :as event-utils]
     [sixsq.nuvla.server.util.time :as time]))
 
 
@@ -117,6 +121,37 @@ passwords) or other services (e.g. TLS credentials for Docker). Creating new
     (str "cannot transform credential-template document to template for subtype: '"
          (:subtype resource) "'")))
 
+;;
+;; actions
+;;
+
+(defn create-job
+  [{{uuid :uuid} :params :as request}]
+  (try
+    (let [id (str resource-type "/" uuid)]
+      (if-let [user-id (auth/current-user-id request)]
+        (let [job-type "credential_check"
+              {{job-id     :resource-id
+                job-status :status} :body} (job/create-job id job-type
+                                                           {:owners   ["group/nuvla-admin"]
+                                                            :view-acl [user-id]}
+                                                           :priority 50)
+              job-msg  (str "starting " id " with async " job-id)]
+          (when (not= job-status 201)
+            (throw (r/ex-response (format "unable to create async job to % log" job-type) 500 id)))
+          (r/map-response job-msg 202 id job-id))
+        (throw (r/ex-response "current authentication has no session identifier" 500 id))))
+    (catch Exception e
+      (or (ex-data e) (throw e)))))
+
+
+(defmethod crud/do-action [resource-type "check"]
+  [{{uuid :uuid} :params :as request}]
+  (let [id       (str resource-type "/" uuid)
+        resource (crud/retrieve-by-id-as-admin id)]
+    (a/throw-cannot-manage resource request)
+    (create-job request)))
+
 
 ;;
 ;; CRUD operations
@@ -173,6 +208,7 @@ passwords) or other services (e.g. TLS credentials for Docker). Creating new
 (defmethod crud/add resource-type
   [{:keys [body] :as request}]
   (let [authn-info (auth/current-authentication request)
+        user-id    (auth/current-user-id request)
         desc-attrs (u/select-desc-keys body)
         [create-resp {:keys [id] :as body}]
         (-> body
@@ -182,11 +218,23 @@ passwords) or other services (e.g. TLS credentials for Docker). Creating new
             (update-in [:template] merge desc-attrs)        ;; ensure desc attrs are validated
             crud/validate
             :template
-            (tpl->credential request))]
-    (-> request
-        (assoc :id id :body (merge body desc-attrs))
-        add-impl
-        (update-in [:body] merge create-resp))))
+            (tpl->credential request))
+
+        response   (-> request
+                       (assoc :id id :body (merge body desc-attrs))
+                       add-impl
+                       (update-in [:body] merge create-resp))
+
+        id         (:resource-id (:body response))
+
+        acl        {:owners   ["group/nuvla-admin"]
+                    :view-acl [user-id]}]
+
+    (when (= (:method body) swarm-tpl/method)
+      (create-job {:params      {:uuid          (u/id->uuid id)
+                                 :resource-name resource-type}
+                   :nuvla/authn auth/internal-identity}))
+    response))
 
 
 (defn create-credential